requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
python-telegram-bot==21.10
together==1.3.14
